Change Interface library in Public lib #84
