+++ libs/LibsClient/CreateEntitiesFunctions/CMakeLists.txt
@@ -4,14 +4,14 @@ set(TARGET ${PROJECT_NAME}_createEntitiesFunctions)
 
 set(SRC
     ${INC_ROOT}/CreateEnemyProjectile.hpp
-    ${INC_ROOT}/CreateEnemyProjectile.cpp
     ${INC_ROOT}/CreateAlliedProjectile.hpp
-    ${INC_ROOT}/CreateAlliedProjectile.cpp
     ${INC_ROOT}/CreatePlayer.hpp
     ${INC_ROOT}/CreateProjectile.hpp
+    ${INC_ROOT}/CreateEnemy.hpp
+    ${INC_ROOT}/CreateObstacle.hpp
 )
 
-add_library(${TARGET} SHARED ${SRC})
+add_library(${TARGET} ${SRC})
 
 target_include_directories(${TARGET}
     PUBLIC ${INC_ROOT}
