+++ libs/LibsClient/CreateEntitiesFunctions/CreateAlliedProjectile.hpp
@@ -9,13 +9,13 @@
 #define CREATEALLYPROJECTILE_HPP_
 
 #include "World/World.hpp"
+#include "GameComponents/AlliedProjectileComponent.hpp"
 #include "GameComponents/CollidableComponent.hpp"
 #include "GameComponents/DamageComponent.hpp"
 #include "GameComponents/DamageRadiusComponent.hpp"
 #include "GameComponents/LifeComponent.hpp"
 #include "GameComponents/LifeTimeComponent.hpp"
 #include "GameComponents/PositionComponent.hpp"
-#include "GameComponents/AlliedProjectileComponent.hpp"
 #include "GameComponents/SizeComponent.hpp"
 #include "GameComponents/VelocityComponent.hpp"
 #include "GameComponents/WeightComponent.hpp"
@@ -26,6 +26,24 @@ namespace ecs
     /// @param world The world in that the Enemy must be created
     /// @param ally Entity who fired an ally projectile
     /// @return Id in size_t of the new Entity
-    std::size_t createNewAlliedProjectile(World &world, Entity &ally);
+    inline std::size_t createNewAlliedProjectile(World &world, Entity &ally)
+    {
+        Position pos = ally.getComponent<Position>();
+        Damage damage = ally.getComponent<Damage>();
+        Velocity velocity = ally.getComponent<Velocity>();
+
+        return world.addEntity()
+            .addComponent<Position>(pos.x, pos.y)
+            .addComponent<Velocity>(velocity.multiplierAbscissa, velocity.multiplierOrdinate)
+            .addComponent<Weight>(1)
+            .addComponent<Size>(2, 1)
+            .addComponent<Lifetime>()
+            .addComponent<Life>(10)
+            .addComponent<Damage>(damage)
+            .addComponent<DamageRadius>(5)
+            .addComponent<Collidable>()
+            .addComponent<AlliedProjectile>()
+            .getId();
+    }
 } // namespace ecs
 #endif /* !CREATEALLYPROJECTILE_HPP_ */
