+++ libs/LibsClient/CreateEntitiesFunctions/CreateEnemy.hpp
@@ -8,6 +8,7 @@
 #ifndef CREATEENEMY_HPP_
 #define CREATEENEMY_HPP_
 
+#include "World/World.hpp"
 #include "GameComponents/CollidableComponent.hpp"
 #include "GameComponents/DamageComponent.hpp"
 #include "GameComponents/DamageRadiusComponent.hpp"
@@ -18,7 +19,6 @@
 #include "GameComponents/SizeComponent.hpp"
 #include "GameComponents/VelocityComponent.hpp"
 #include "GameComponents/WeightComponent.hpp"
-#include "World/World.hpp"
 
 namespace ecs
 {
