+++ libs/LibsClient/CreateEntitiesFunctions/CreateProjectile.hpp
@@ -8,6 +8,7 @@
 #ifndef CREATEPROJECTILE_HPP_
 #define CREATEPROJECTILE_HPP_
 
+#include "World/World.hpp"
 #include "GameComponents/CollidableComponent.hpp"
 #include "GameComponents/DamageComponent.hpp"
 #include "GameComponents/DamageRadiusComponent.hpp"
@@ -18,7 +19,6 @@
 #include "GameComponents/SizeComponent.hpp"
 #include "GameComponents/VelocityComponent.hpp"
 #include "GameComponents/WeightComponent.hpp"
-#include "World/World.hpp"
 
 namespace ecs
 {
